life-agent/requirements.txt
# Core Dependencies
python-telegram-bot==20.7
openai==1.12.0
anthropic==0.18.1
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Browser Automation
playwright==1.41.0
beautifulsoup4==4.12.3
requests==2.31.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
pillow==10.2.0

# Utilities
pytz==2024.1
python-dateutil==2.8.2
pyyaml==6.0.1
schedule==1.2.0

# Financial
plaid-python==16.0.0

# Voice (optional)
openai-whisper==20231117
pydub==0.25.1

# Calendar
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Image Processing
opencv-python==4.9.0.80
pytesseract==0.3.10
